Conflicts:
	VERSION.pcp
	debian/changelog
	src/libpcp/src/derive.c
	src/libpcp/src/p_result.c
	src/libpcp/src/util.c
	src/pmcd/rc_pcp
	src/pmcd/src/pmcd.c

Mostly version and changelog conflicts.  Some minor code changes in the
area of printf formatting, and forward port the restartcond change to
track the rc script split pcp -> pmcd & pmlogger.

+++ src/libpcp/src/logconnect.c
@@ -210,7 +210,7 @@ __pmConnectLogger(const char *hostname, int *pid, int *port)
 	    handshake[0].c_vala = LOG_PDU_VERSION;
 	    handshake[0].c_valb = 0;
 	    handshake[0].c_valc = 0;
-	    sts = __pmSendCreds(fd, getpid(), 1, handshake);
+	    sts = __pmSendCreds(fd, (int)getpid(), 1, handshake);
 	}
 	else
 	    sts = PM_ERR_IPC;
